+++ FlashDevelop/MainForm.cs
@@ -3774,7 +3774,7 @@ namespace FlashDevelop
                 if (position > -1)
                 {
                     var message = TextHelper.GetString("Info.RunningProcess");
-                    TraceManager.Add(message + " " + args.Substring(0, position) + " " + args.Substring(position + 1), (int)TraceType.ProcessStart);
+                    TraceManager.Add(message + " " + args.Substring(0, position) + " " + args.Substring(position + 1), TraceType.ProcessStart);
                     var psi = new ProcessStartInfo
                     {
                         WorkingDirectory = WorkingDirectory,
@@ -3821,11 +3821,11 @@ namespace FlashDevelop
                 if (position < 0)
                 {
                     string message = TextHelper.GetString("Info.NotEnoughArguments");
-                    TraceManager.Add(message + " " + args, (int)TraceType.Error);
+                    TraceManager.Add(message + " " + args, TraceType.Error);
                     return;
                 }
                 string message2 = TextHelper.GetString("Info.RunningProcess");
-                TraceManager.Add(message2 + " " + args.Substring(0, position) + " " + args.Substring(position + 1), (int)TraceType.ProcessStart);
+                TraceManager.Add(message2 + " " + args.Substring(0, position) + " " + args.Substring(position + 1), TraceType.ProcessStart);
                 processRunner.Run(args.Substring(0, position), args.Substring(position + 1));
                 ButtonManager.UpdateFlaggedButtons();
             }
@@ -3838,12 +3838,12 @@ namespace FlashDevelop
         /// <summary>
         /// Handles the incoming info output
         /// </summary>
-        static void ProcessOutput(object sender, string line) => TraceManager.AddAsync(line, (int)TraceType.Info);
+        static void ProcessOutput(object sender, string line) => TraceManager.AddAsync(line, TraceType.Info);
 
         /// <summary>
         /// Handles the incoming error output
         /// </summary> 
-        static void ProcessError(object sender, string line) => TraceManager.AddAsync(line, (int)TraceType.ProcessError);
+        static void ProcessError(object sender, string line) => TraceManager.AddAsync(line, TraceType.ProcessError);
 
         /// <summary>
         /// Handles the ending of a process
@@ -3854,7 +3854,7 @@ namespace FlashDevelop
             else
             {
                 var result = $"Done({exitCode})";
-                TraceManager.Add(result, (int)TraceType.ProcessEnd);
+                TraceManager.Add(result, TraceType.ProcessEnd);
                 var te = new TextEvent(EventType.ProcessEnd, result);
                 EventManager.DispatchEvent(this, te);
                 ButtonManager.UpdateFlaggedButtons();
